requirements.txt
absl-py==0.11.0
allure-pytest==2.8.18
allure-python-commons==2.8.18
altgraph==0.17
Appium-Python-Client==1.0.2
arrow==0.17.0
astunparse==1.6.3
atomicwrites==1.4.0
attrs==20.2.0
baidu-aip==2.2.18.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
cx-Oracle==8.0.1
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.34.0
h5py==2.10.0
idna==2.10
iniconfig==1.1.1
Keras-Preprocessing==1.1.2
Markdown==3.3.3
more-itertools==8.6.0
Naked==0.1.31
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==20.4
pefile==2019.4.18
Pillow==8.0.1
pluggy==0.13.1
protobuf==3.14.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.9
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.11
PyMySQL==0.10.1
pyparsing==2.4.7
pytest==6.0.2
python-dateutil==2.8.1
pywin32-ctypes==0.2.0
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
selenium==3.141.0
shellescape==3.8.1
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
